完善后台管理基本功能

+++ blog/src/main/java/com/beifeng/service/impl/TypeServiceImpl.java
@@ -3,6 +3,8 @@ package com.beifeng.service.impl;
 import com.beifeng.dao.TypeMapper;
 import com.beifeng.domain.Type;
 import com.beifeng.service.TypeService;
+import com.beifeng.util.DateTimeUtil;
+import com.beifeng.util.UUIDUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,12 +25,16 @@ public class TypeServiceImpl implements TypeService {
     @Transactional
     @Override
     public Integer saveType(Type type) {
+        type.setId(UUIDUtil.getUUID());
+        type.setCreateTime(DateTimeUtil.getSysTime());
+        type.setUpdateTime(DateTimeUtil.getSysTime());
+
         return typeMapper.saveType(type);
     }
 
     @Transactional
     @Override
-    public Type getType(Long id) {
+    public Type getType(String id) {
         return typeMapper.getTypeById(id);
     }
 
@@ -40,9 +46,14 @@ public class TypeServiceImpl implements TypeService {
 
     @Transactional
     @Override
-    public String updateType(String name,Long id) {
+    public String updateType(String name,String id) {
         String msg;
-        Integer count = typeMapper.updateType(name,id);
+
+        Type type = new Type();
+        type.setId(id);
+        type.setName(name);
+        type.setUpdateTime(DateTimeUtil.getSysTime());
+        Integer count = typeMapper.updateType(type);
 
         if (count == 0 ) {
             msg = "修改失败";
@@ -55,9 +66,21 @@ public class TypeServiceImpl implements TypeService {
 
     @Transactional
     @Override
-    public String deleteType(Long id) {
-        Integer count = typeMapper.deleteType(id);
+    public String deleteType(String id) {
+
         String msg;
+        /*
+            查询分类下有多少博客，
+                博客为0，可直接删除，
+                博客不为0，不可删除此分类
+        */
+        Integer blogCount = typeMapper.getBlogCountByTypeId(id);
+        if (blogCount>0){
+            msg = "请先删除分类下所有博客";
+            return msg;
+        }
+
+        Integer count = typeMapper.deleteType(id);
         if (count == 0 ) {
             msg = "删除失败";
         } else {
